<utmpx.h>: Change struct utmpx's ut_type from uint16_t to short, per POSIX.
See https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/utmpx.h.html

Our valid ut_types are 0-8 currently, so there should not be any issues.

Also fix a no longer necessary cast in utmp_update.

+++ libexec/utmp_update/utmp_update.c
@@ -85,7 +85,7 @@ main(int argc, char *argv[])
 	case DEAD_PROCESS:
 		break;
 	default:
-		errx(1, "Invalid utmpx type %d", (int)utx->ut_type);
+		errx(1, "Invalid utmpx type %hd", utx->ut_type);
 	}
 
 	if (ruid != 0) {
